refactor : exception 처리 분리
