+++ src/main/java/nextstep/subway/member/application/MemberService.java
@@ -1,14 +1,16 @@
 package nextstep.subway.member.application;
 
+import org.springframework.stereotype.Service;
+
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import nextstep.subway.member.exception.MemberNotFoundException;
 
 @Service
 public class MemberService {
+    public static final String ERROR_MEMBER_NOT_FOUND = "사용자를 찾을 수 없습니다.";
     private MemberRepository memberRepository;
 
     public MemberService(MemberRepository memberRepository) {
@@ -21,16 +23,21 @@ public class MemberService {
     }
 
     public MemberResponse findMember(Long id) {
-        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        Member member = findById(id);
         return MemberResponse.of(member);
     }
 
     public void updateMember(Long id, MemberRequest param) {
-        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        Member member = findById(id);
         member.update(param.toMember());
     }
 
     public void deleteMember(Long id) {
         memberRepository.deleteById(id);
+    }
+
+    public Member findById(Long id) {
+        return memberRepository.findById(id)
+            .orElseThrow(() -> new MemberNotFoundException(ERROR_MEMBER_NOT_FOUND));
     }
 }
