+++ src/main/java/nextstep/subway/member/exception/MemberNotFoundException.java
@@ -1,0 +1,16 @@
+package nextstep.subway.member.exception;
+
+import javax.persistence.EntityNotFoundException;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class MemberNotFoundException extends EntityNotFoundException {
+    public MemberNotFoundException() {
+    }
+
+    public MemberNotFoundException(String message) {
+        super(message);
+    }
+}
